Replace REAL*8 with real(real64) throughout codebase; add HITSUR typed dispatch
Replaced all REAL*8 declarations with real(real64) from iso_fortran_env
across 133 source files. Added use iso_fortran_env, only: real64 to each
affected compilation unit, including subroutines without IMPLICIT NONE
(ASTRO1, MATH, SVD, LDM12) that the automated pass initially missed.

Also added top-level polymorphic surface dispatch in HITSUR (RAYTRA2.f90):
for any surface type extending surface_type, HITSUR now delegates
intersection to ldm%surfaces(R_I)%s%intersect() and hands off physics
to INTERACK, without modifying HITASP/HITFLA. New surface types only need
to implement intersect/real_trace/paraxial_trace to be picked up
automatically. A stale-data guard (ubound check) prevents wrong typed
surfaces from being used mid-macro before load_surfaces_from_alens runs.

+++ src/BICUBIC.f90
@@ -1,8 +1,9 @@
 SUBROUTINE BICUBIC(X1,X2,X3,X4,Y1,Y2,F1,F2,F3,F4 &
 &,X,Y,F)
+   use iso_fortran_env, only: real64
    IMPLICIT NONE
-   REAL*8 X1,X2,X3,X4,Y1,Y2,F1,F2,F3,F4,X,Y,F
-   REAL*8 M1X,M2X,FX1,FX2,MY
+   real(real64) X1,X2,X3,X4,Y1,Y2,F1,F2,F3,F4,X,Y,F
+   real(real64) M1X,M2X,FX1,FX2,MY
 !       X DIRECTION
    M1X=(F2-F1)/(X2-X1)
    M2X=(F4-F3)/(X4-X3)
